+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Unreleased
+- Added ControllerTrait.
+- Deprecated AbstractController.
 
 ## [2.1.0] - 2019-06-13
 - Added ActionResult.